lekce_1/pyproject.toml
[project]
name = "4-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.21.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "yfinance>=0.2.66",
]
